feat: se actualizan roles y se agrega segurar al seed

+++ src/seed/seed.service.ts
@@ -2,16 +2,21 @@ import {
     BadRequestException,
     Injectable,
     InternalServerErrorException,
+    UnauthorizedException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { CONTENT_TYPES, DAYS, ROLES } from './data';
+import { CONTENT_TYPES, DAYS, FIRST_USER, ROLES } from './data';
 import { Role } from 'src/role/entities/role.entity';
 import { Day } from 'src/day/entities/day.entity';
 import { ContentType } from 'src/content-type/entities/content-type.entity';
+import { ConfigService } from '@nestjs/config';
+import { UserService } from 'src/user/user.service';
+import { CreateUserDto } from 'src/user/dto/create-user.dto';
 
 @Injectable()
 export class SeedService {
+    private seedSecret: string;
     constructor(
         @InjectModel(Day.name)
         private readonly dayModel: Model<Day>,
@@ -19,8 +24,14 @@ export class SeedService {
         private readonly contentTypeModel: Model<ContentType>,
         @InjectModel(Role.name)
         private readonly roleModel: Model<Role>,
-    ) {}
-    async executeSeed() {
+        private readonly configService: ConfigService,
+        private readonly userService: UserService,
+    ) {
+        this.seedSecret = this.configService.get<string>('seedSecret');
+    }
+    async executeSeed(password: string) {
+        if (password !== this.seedSecret)
+            throw new UnauthorizedException('Sin autorización.');
         try {
             await this.dayModel.deleteMany({});
             await this.contentTypeModel.deleteMany({});
@@ -28,6 +39,16 @@ export class SeedService {
             await this.dayModel.insertMany(DAYS);
             await this.contentTypeModel.insertMany(CONTENT_TYPES);
             await this.roleModel.insertMany(ROLES);
+            const role: Role = await this.roleModel.findOne({
+                name: ROLES[0].name,
+            });
+            if (!role)
+                throw new InternalServerErrorException('No se encontro el Rol');
+            const createUserDto: CreateUserDto = {
+                ...FIRST_USER,
+                roles: [role._id],
+            };
+            await this.userService.create(createUserDto);
             return 'Seed Executed';
         } catch (error) {
             if (error === 'E11000') {
@@ -35,7 +56,6 @@ export class SeedService {
                     'Algun campo duplicado Duplicado',
                 );
             }
-            console.log(error);
             throw new InternalServerErrorException(
                 'No se puede ejecutar el Seed',
             );
